# perf: prefetch span attributes during ingest and sampling