--- a/sample/sample.go
+++ b/sample/sample.go
@@ -15,7 +15,7 @@ import (
 
 type Sampler interface {
 	GetSampleRate(trace *types.Trace) (rate uint, keep bool, reason string, key string)
-	GetKeyFields() []string
+	GetKeyFields() ([]string, []string)
 	Start() error
 	Stop()
 }
@@ -164,4 +164,30 @@ func (d *dynsamplerMetricsRecorder) RecordMetrics(sampler dynsampler.Sampler, ke
 	}
 	d.met.Histogram(d.prefix+"_sampler_key_cardinality", float64(numTraceKey))
 	d.met.Histogram(d.prefix+"_sample_rate", float64(rate))
+}
+
+// getKeyFields returns the fields that should be used as keys for the sampler.
+// It returns two slices: the first contains fields with the root prefix removed,
+// and the second contains fields that do not have the root prefix.
+func getKeyFields(fields []string) ([]string, []string) {
+	if len(fields) == 0 {
+		return nil, nil
+	}
+
+	new := make([]string, 0, len(fields))
+	nonRootFields := make([]string, 0, len(fields))
+
+	for _, field := range fields {
+		if strings.HasPrefix(field, RootPrefix) {
+			new = append(new, field[len(RootPrefix):])
+		} else {
+			nonRootFields = append(nonRootFields, field)
+		}
+	}
+
+	if len(new) == 0 {
+		return nonRootFields, nonRootFields
+	}
+
+	return append(new, nonRootFields...), nonRootFields
 }