--- a/sample/trace_key.go
+++ b/sample/trace_key.go
@@ -3,7 +3,6 @@ package sample
 import (
 	"bytes"
 	"fmt"
-	"slices"
 	"sort"
 	"strconv"
 	"strings"
@@ -21,7 +20,6 @@ const maxKeyLength = 100
 type traceKey struct {
 	fields         []string
 	rootOnlyFields []string
-	allFields      []string // all fields, including root-only fields
 	useTraceLength bool
 	keyBuilder     *bytes.Buffer
 	distinctValue  *distinctValue
@@ -44,7 +42,6 @@ func newTraceKey(fields []string, useTraceLength bool) *traceKey {
 	return &traceKey{
 		fields:         nonRootFields,
 		rootOnlyFields: rootOnlyFields,
-		allFields:      slices.Concat(nonRootFields, rootOnlyFields),
 		useTraceLength: useTraceLength,
 		distinctValue:  &distinctValue{buf: make([]byte, 0, 1024)},
 		keyBuilder:     &bytes.Buffer{},
@@ -58,14 +55,6 @@ func (d *traceKey) build(trace *types.Trace) (string, int) {
 
 	// for each field, for each span, get the value of that field
 	spans := trace.GetSpans()
-	for _, s := range spans {
-		if s.IsRoot {
-			s.Data.MemoizeFields(d.allFields...)
-			continue
-		}
-		s.Data.MemoizeFields(d.fields...)
-	}
-
 	d.distinctValue.Reset(d.fields, maxKeyLength)
 	d.keyBuilder.Reset()
 outer: