## Which problem is this PR solving?

Fetch multiple data from the event payload is faster than doing it one at a time

## Short description of the changes

TODO: I need to add tests for the changes