@ianwilkes 2025-06-27
I decided to try going further and put our core metadata fields in the struct as first-class fields. This is reasonably ergonomic from the outside, but very verbose on the inside, so I have mixed feelings about it. It generates fewer allocations per field, but ultimately memoizes more fields than are used in the route benchmark, so it looks like a wash on memory but should do better in production, I think. It's definitely faster vs performance_work_branch:

```
benchmark                old ns/op     new ns/op     delta
BenchmarkRouterBatch     44137         18956         -57.05%

benchmark                old allocs     new allocs     delta
BenchmarkRouterBatch     61             62             +1.64%

benchmark                old bytes     new bytes     delta
BenchmarkRouterBatch     65020         66156         +1.75%
```

In theory, this should make life easier and faster for the collector, but I'm open to the idea that maybe this is more complexity than is warranted and we should just stick with memoization. Profiling shows that a lot of of allocations here are unrelated to this specific code; there's some general cleanup we need to do in route.go. Most of the CPU time is just reading all the bytes, and the messsagepack parse; there may be more improvement possible on both those fronts.

I'm not too confident in this list of fields either. I asked claude for advice and here's what I got, this may be a good starting point if we want to go further with this:

Fields that are commonly used but NOT in the dedicated list:

1. "meta.refinery.original_sample_rate" - Used frequently in collect.go to preserve the original sample rate
2. "meta.span_count" - Used 6 times in collect.go for root span enrichment
3. "meta.event_count" - Used 3 times for root span enrichment
4. "meta.span_event_count" - Used 3 times for root span enrichment
5. "meta.span_link_count" - Used 3 times for root span enrichment
6. "meta.refinery.reason" - Used 3 times to record sampling decisions
7. "meta.refinery.local_hostname" - Used 3 times for debugging/tracking

Recommendations:

1. Consider adding meta.refinery.original_sample_rate to the dedicated fields list since it's used on every sampled span and is important for rate calculations.
2. Maybe add the count fields (meta.span_count, meta.event_count, etc.) if AddCountsToRoot is commonly enabled, as these would be accessed frequently on root spans.
3. Keep as-is the other fields like meta.refinery.reason and meta.refinery.local_hostname since they're used less frequently and only in specific scenarios.