@ianwilkes 2025-06-28
Taking this a bit further to only do a single pass on msgpack data, in the ideal batch/msgpack case. I'd love to see how this performs in prod.

```
benchmark                old ns/op     new ns/op     delta
BenchmarkRouterBatch     44137         12024         -72.76%
```